+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2022-10-11
+### Platform support
+- Linux
+
+### Fixes
+- mode change configurations affecting buffering corrected
+
+### Changes
+
+### Removals
+
+### Additions
+
+### Documentation
+- Simple demo application added to show API usage
+
 ## [0.1.0] - 2022-10-01
 ### Platform support
 - work in progress